requirements.txt
flask
flask-cors
pillow
numpy
# Optional (for future real-model integration):
# opencv-python-headless
# torch
# torchvision
